Alright, some big changes on this one.
I removed the two subpython scripts in favor of one that condenses both operations of manipulating the Pfsense JSON config that is retrieved. This will make things faster and a little less cluttered in my opinion.

I also made the bash script check for the nord-vpn-server-find directory and pull it if it did not exist. Should make initial setup easier until I can poll the nord api directly from the python script. And who knows, perhaps this can be all python at some point. Wouldn't that be a wonder.